pyproject.toml
[tool.poetry]
name = "pythonproject5"
version = "0.1.0"
description = ""
authors = ["Owwwen86 <smirnovkirill@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
envsubst = "^0.1.5"
djangorestframework = "^3.14.0"
chardet = "^5.1.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.1"
pydantic = "^1.10.7"
requests = "^2.28.2"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"


[tool.poetry.group.dev.dependencies]
ansible-vault-win = "^1.0.0"
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#[tool.pytest.ini_options]
#DJANGO_SETTINGS_MODULE = 'todolist.settings'
#python_files = [
#    'tests.py',
#    'test_*.py',
#    '*_tests.py'
#]
